Disable nextjs built-in localization routing
Nextjs localization does not work with SSG. In this commit I disabled
all configurations for the buil-in localization, and added a locale subpath
for all pages.

This is quite verbose, and forces us to have a lot of boilerplate (getStaticProps
and getStaticPaths for all pages). It also forces us to have the locale in the URL
even if it is the default one.

I added a redirect in the 404 that tries to redirect to the correct locale
based on browser preferences.